Improve the logging of errors in the compress/decompress tasks
Rename the zip archive task to ZipFileExtractToDirectory
