+++ src/Microsoft.DotNet.Build.Tasks/PackageFiles/codeOptimization.targets
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="12.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <UsingTask TaskName="ExtractZipFile" AssemblyFile="$(BuildToolsTaskDir)Microsoft.DotNet.Build.Tasks.dll"/>
+  <UsingTask TaskName="ZipFileExtractToDirectory" AssemblyFile="$(BuildToolsTaskDir)Microsoft.DotNet.Build.Tasks.dll"/>
 
   <PropertyGroup>
     <OptimizationDataVersion>1.0.0-prerelease</OptimizationDataVersion>
@@ -20,7 +20,7 @@
     </PropertyGroup>
 
     <!-- Extract the profile data from the archive -->
-    <ExtractZipFile SourceArchive="$(OptimizationDataDir)/$(AssemblyName).zip"
+    <ZipFileExtractToDirectory SourceArchive="$(OptimizationDataDir)/$(AssemblyName).zip"
                     DestinationDirectory="$(OptimizationDataForAssemblyDir)"
                     OverwriteDestination="true" />
